# docs/openapi/evidsight-v1.yaml
openapi: 3.1.0
info:
  title: EvidSight 外部 API v1
  version: 1.0.0
  description: |
    EvidSight Knowledge 与 Research 服务供 Web 消费的外部 v1 契约。本文件是 Auth、
    Knowledge Base、Document、Chat、Conversation、Research、Evidence 与 Report 请求、响应
    和 SSE data 字段与约束的唯一权威来源；API.md 定义行为、
    权限、状态码与兼容语义，本文只定义 Schema。新增或修改字段必须先修订本文件，
    再经 FastAPI 路由清单 ↔ OpenAPI 路径清单 ↔ Provider 契约测试三方一致校验
    （API.md §3.1 / ROADMAP M2 退出门禁）。
servers:
  - url: ""
paths:
...
  /api/v1/research/tasks:
    post:
      tags: [Research v1]
      operationId: createResearchTaskV1
      summary: 幂等创建研究任务
      parameters:
        - {name: Idempotency-Key, in: header, required: true, schema: {type: string, minLength: 1, maxLength: 128}}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ResearchTaskCreate"}
      responses:
        "202":
          description: 任务已接受
          content: {application/json: {schema: {$ref: "#/components/schemas/ResearchTaskCreateResponse"}}}
        "409": {$ref: "#/components/responses/Conflict"}
        "422": {$ref: "#/components/responses/ValidationError"}
    get:
      tags: [Research v1]
      operationId: listResearchTasksV1
      summary: 研究任务列表
      parameters:
        - {$ref: "#/components/parameters/PageParam"}
        - {$ref: "#/components/parameters/PageSizeParam"}
        - {name: status, in: query, schema: {type: string}}
        - {name: keyword, in: query, schema: {type: string}}
      responses:
        "200":
          description: 研究任务分页列表
          content: {application/json: {schema: {$ref: "#/components/schemas/ResearchTaskList"}}}
  /api/v1/research/tasks/{task_id}:
    parameters:
      - {$ref: "#/components/parameters/TaskIdParam"}
    get:
      tags: [Research v1]
      operationId: getResearchTaskV1
      summary: 研究任务详情
      responses:
        "200":
          description: 任务详情
          content: {application/json: {schema: {$ref: "#/components/schemas/ResearchTask"}}}
        "404": {$ref: "#/components/responses/NotFound"}
    delete:
      tags: [Research v1]
      operationId: deleteResearchTaskV1
      summary: 删除研究任务
      responses:
        "204": {description: 已删除，无正文}
        "404": {$ref: "#/components/responses/NotFound"}
  /api/v1/research/tasks/{task_id}/cancel:
    parameters: [{$ref: "#/components/parameters/TaskIdParam"}]
    post:
      tags: [Research v1]
      operationId: cancelResearchTaskV1
      summary: 请求取消研究任务
      responses:
        "202": {description: 取消已接受, content: {application/json: {schema: {$ref: "#/components/schemas/ResearchTask"}}}}
        "404": {$ref: "#/components/responses/NotFound"}
  /api/v1/research/tasks/{task_id}/resume:
    parameters: [{$ref: "#/components/parameters/TaskIdParam"}]
    post:
      tags: [Research v1]
      operationId: resumeResearchTaskV1
      summary: 断点续跑研究任务
      responses:
        "202": {description: 恢复已接受, content: {application/json: {schema: {$ref: "#/components/schemas/ResearchTask"}}}}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
  /api/v1/research/tasks/{task_id}/state:
    parameters: [{$ref: "#/components/parameters/TaskIdParam"}]
    get:
      tags: [Research v1]
      operationId: getResearchTaskStateV1
      summary: 研究状态轮询快照
      responses:
        "200": {description: 状态快照, content: {application/json: {schema: {$ref: "#/components/schemas/ResearchTaskState"}}}}
        "404": {$ref: "#/components/responses/NotFound"}
  /api/v1/research/tasks/{task_id}/events:
    parameters: [{$ref: "#/components/parameters/TaskIdParam"}]
    get:
      tags: [Research v1]
      operationId: streamResearchTaskEventsV1
      summary: 订阅研究任务事件
      parameters:
        - {name: Last-Event-ID, in: header, required: false, schema: {type: string}}
      responses:
        "200":
          description: canonical Research SSE
          content:
            text/event-stream:
              schema: {type: string}
              x-sse-data-schema: {$ref: "#/components/schemas/ResearchEventData"}
        "404": {$ref: "#/components/responses/NotFound"}
  /api/v1/research/tasks/{task_id}/report:
    parameters: [{$ref: "#/components/parameters/TaskIdParam"}]
    get:
      deprecated: true
      tags: [Research v1]
      operationId: getResearchTaskReportCompatibilityV1
      summary: 迁移兼容任务报告读取
      responses:
        "200": {description: 兼容报告对象}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
  /api/v1/research/tasks/{task_id}/evidence:
    parameters: [{$ref: "#/components/parameters/TaskIdParam"}]
    get:
      tags: [Evidence v1]
      operationId: listResearchTaskEvidenceV1
      summary: 任务 Evidence 列表
      responses:
        "200": {description: Evidence 列表, content: {application/json: {schema: {$ref: "#/components/schemas/EvidenceList"}}}}
        "404": {$ref: "#/components/responses/NotFound"}
  /api/v1/evidence/{evidence_id}:
    parameters:
      - {name: evidence_id, in: path, required: true, schema: {type: string, format: uuid}}
    get:
      tags: [Evidence v1]
      operationId: getEvidenceV1
      summary: Evidence 详情
      responses:
        "200": {description: Evidence 详情, content: {application/json: {schema: {$ref: "#/components/schemas/Evidence"}}}}
        "404": {$ref: "#/components/responses/NotFound"}
  /api/v1/evidence/{evidence_id}/relations:
    parameters:
      - {name: evidence_id, in: path, required: true, schema: {type: string, format: uuid}}
    get:
      tags: [Evidence v1]
      operationId: listEvidenceRelationsV1
      summary: Evidence 关系
      responses:
        "200": {description: 关系列表, content: {application/json: {schema: {$ref: "#/components/schemas/EvidenceRelationList"}}}}
        "404": {$ref: "#/components/responses/NotFound"}
  /api/v1/reports/{report_id}:
    parameters:
      - {$ref: "#/components/parameters/ReportIdParam"}
    get:
      tags: [Report v1]
      operationId: getReportV1
      summary: 正式报告详情
      responses:
        "200": {description: published Report, content: {application/json: {schema: {$ref: "#/components/schemas/Report"}}}}
        "404": {$ref: "#/components/responses/NotFound"}
  /api/v1/reports/{report_id}/sections/{section_id}:
    parameters:
      - {$ref: "#/components/parameters/ReportIdParam"}
      - {name: section_id, in: path, required: true, schema: {type: string, format: uuid}}
    get:
      tags: [Report v1]
      operationId: getReportSectionV1
      summary: 正式报告章节
      responses:
        "200": {description: Report 章节, content: {application/json: {schema: {$ref: "#/components/schemas/ReportSection"}}}}
        "404": {$ref: "#/components/responses/NotFound"}
components:
  parameters:
    PageParam:
      name: page
      in: query
      required: false
      description: 页码，从 1 开始
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      name: page_size
      in: query
      required: false
      description: 每页条数
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    TaskIdParam:
      name: task_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ReportIdParam:
      name: report_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    NotFound:
      description: 资源不存在或不可见
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Conflict:
      description: 状态或名称冲突
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    ValidationError:
      description: 请求参数校验失败（E9003）
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    ErrorEnvelope:
      type: object
      description: API.md §4 标准错误信封
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/Error"
      additionalProperties: false
    Error:
      type: object
      required: [error_code, message, request_id, retryable, details]
      properties:
        error_code:
          type: string
          pattern: "^[A-Z][A-Z0-9_]+$"
        message:
          type: string
        request_id:
          type: string
        retryable:
          type: boolean
        details:
          type: object
          additionalProperties: true
          description: 安全、结构化、可操作信息；不含正文、密码、Token、SQL、路径或堆栈
      additionalProperties: false
    # ---- Knowledge Base ----
...
    StreamErrorEventData:
      type: object
      required: [error_code, message, retryable]
      properties:
        error_code: {type: string}
        message: {type: string}
        retryable: {type: boolean}
...
    ResearchRequirements:
      type: object
      required: [task_type]
      additionalProperties: false
      properties:
        task_type: {type: string, enum: [comparison, explainer, analysis]}
        depth: {type: string, enum: [quick], default: quick}
        max_sources: {type: integer, minimum: 1, maximum: 50, default: 10}
        language: {type: string, minLength: 2, maxLength: 10, default: zh}
    ResearchTaskCreate:
      type: object
      required: [topic, requirements]
      additionalProperties: false
      properties:
        topic: {type: string, minLength: 1, maxLength: 500}
        requirements: {$ref: "#/components/schemas/ResearchRequirements"}
        source_strategy: {type: string, enum: [knowledge, web, hybrid], default: web}
        knowledge_base_ids:
          type: array
          maxItems: 50
          items: {type: string, format: uuid}
          default: []
    ResearchTaskCreateResponse:
      type: object
      required: [task_id, status, created_at, direct_answer, idempotent_replayed, report_id]
      properties:
        task_id: {type: string, format: uuid}
        status: {type: string}
        created_at: {type: string, format: date-time}
        direct_answer: {type: boolean}
        idempotent_replayed: {type: boolean}
        report_id: {type: [string, "null"], format: uuid}
    ResearchProgress:
      type: object
      required: [completed_steps, total_steps, progress]
      properties:
        completed_steps: {type: integer, minimum: 0}
        total_steps: {type: integer, minimum: 0}
        progress: {type: number, minimum: 0, maximum: 1}
    ResearchTask:
      type: object
      required: [task_id, topic, status, requirements, progress, total_sources, total_evidence, created_at, report_id]
      properties:
        task_id: {type: string, format: uuid}
        topic: {type: string}
        status: {type: string}
        current_phase: {type: [string, "null"]}
        requirements: {$ref: "#/components/schemas/ResearchRequirements"}
        progress: {$ref: "#/components/schemas/ResearchProgress"}
        total_sources: {type: integer, minimum: 0}
        total_evidence: {type: integer, minimum: 0}
        error_code: {type: [string, "null"]}
        error_message: {type: [string, "null"]}
        recoverable: {type: [boolean, "null"]}
        report_id: {type: [string, "null"], format: uuid}
        created_at: {type: string, format: date-time}
        started_at: {type: [string, "null"], format: date-time}
        completed_at: {type: [string, "null"], format: date-time}
    ResearchTaskListItem:
      type: object
      required: [task_id, topic, status, task_type, total_sources, total_evidence, created_at, report_id]
      properties:
        task_id: {type: string, format: uuid}
        topic: {type: string}
        status: {type: string}
        task_type: {type: string}
        total_sources: {type: integer, minimum: 0}
        total_evidence: {type: integer, minimum: 0}
        report_id: {type: [string, "null"], format: uuid}
        created_at: {type: string, format: date-time}
        completed_at: {type: [string, "null"], format: date-time}
    ResearchTaskList:
      type: object
      required: [items, page, page_size, total]
      properties:
        items: {type: array, items: {$ref: "#/components/schemas/ResearchTaskListItem"}}
        page: {type: integer, minimum: 1}
        page_size: {type: integer, minimum: 1}
        total: {type: integer, minimum: 0}
    ResearchTaskState:
      type: object
      required: [task_id, status, progress, steps, created_at, report_id]
      properties:
        task_id: {type: string, format: uuid}
        topic: {type: string}
        status: {type: string}
        current_phase: {type: [string, "null"]}
        progress: {$ref: "#/components/schemas/ResearchProgress"}
        steps: {type: array, items: {type: object}}
        report_id: {type: [string, "null"], format: uuid}
        created_at: {type: string, format: date-time}
        started_at: {type: [string, "null"], format: date-time}
        completed_at: {type: [string, "null"], format: date-time}
    ResearchEventData:
      anyOf:
        - {$ref: "#/components/schemas/ResearchTaskState"}
        - {$ref: "#/components/schemas/StreamErrorEventData"}
        - {type: object, additionalProperties: true}
    # ---- Evidence / Report ----
    Evidence:
      type: object
      required: [evidence_id, task_id, source_type, display_title, validity, used_in_sections]
      properties:
        evidence_id: {type: string, format: uuid}
        task_id: {type: string, format: uuid}
        source_type: {type: string, enum: [internal, web]}
        display_title: {type: [string, "null"]}
        location_summary: {type: [string, "null"]}
        source_observed_at: {type: [string, "null"], format: date-time}
        score_summary: {type: [object, "null"], additionalProperties: true}
        validity: {type: string}
        relevance_score: {type: [number, "null"]}
        used_in_sections: {type: array, items: {type: string}}
        knowledge_base_id: {type: [string, "null"], format: uuid}
        document_id: {type: [string, "null"], format: uuid}
        document_version_id: {type: [string, "null"], format: uuid}
        segment_id: {type: [string, "null"], format: uuid}
        document_display_name_snapshot: {type: [string, "null"]}
        url: {type: [string, "null"], format: uri}
        fetched_at: {type: [string, "null"], format: date-time}
        content: {type: [string, "null"]}
    EvidenceList:
      type: object
      required: [items, total]
      properties:
        items: {type: array, items: {$ref: "#/components/schemas/Evidence"}}
        total: {type: integer, minimum: 0}
    EvidenceRelation:
      type: object
      required: [relation_id, relation_type, confidence, claim]
      properties:
        relation_id: {type: integer}
        relation_type: {type: string, enum: [supports, contradicts, context]}
        confidence: {type: number}
        rationale_summary: {type: [string, "null"]}
        claim: {type: object, additionalProperties: true}
    EvidenceRelationList:
      type: object
      required: [items, total]
      properties:
        items: {type: array, items: {$ref: "#/components/schemas/EvidenceRelation"}}
        total: {type: integer, minimum: 0}
    ReportEvidenceReference:
      type: object
      required: [evidence_id, source_type]
      properties:
        evidence_id: {type: string, format: uuid}
        source_type: {type: string, enum: [internal, web]}
        url: {type: [string, "null"], format: uri}
        fetched_at: {type: [string, "null"], format: date-time}
    ReportSection:
      type: object
      required: [section_id, heading, content, sequence, evidence]
      properties:
        section_id: {type: string, format: uuid}
        heading: {type: string}
        content: {type: string}
        sequence: {type: integer, minimum: 0}
        evidence: {type: array, items: {$ref: "#/components/schemas/ReportEvidenceReference"}}
    Report:
      type: object
      required: [report_id, task_id, revision, status, title, language, sections]
      properties:
        report_id: {type: string, format: uuid}
        task_id: {type: string, format: uuid}
        revision: {type: integer, minimum: 1}
        status: {type: string, enum: [published]}
        title: {type: string}
        language: {type: string}
        evidence_completeness: {type: [string, "null"]}
        limitations_summary: {type: [string, "null"]}
        published_at: {type: [string, "null"], format: date-time}
        sections: {type: array, items: {$ref: "#/components/schemas/ReportSection"}}